+++ src/fides/api/models/client.py
@@ -23,8 +23,9 @@ from fides.api.cryptography.schemas.jwt import (
 from fides.api.db.base_class import Base
 from fides.api.models.fides_user import FidesUser
 from fides.api.oauth.jwt import generate_jwe
-from fides.config import FidesConfig
+from fides.config import CONFIG, FidesConfig
 
+DEFAULT_OAUTH_CLIENT_KEY = "default_oauth_client"
 DEFAULT_SCOPES: list[str] = []
 DEFAULT_ROLES: list[str] = []
 DEFAULT_SYSTEMS: list[str] = []
@@ -109,8 +110,11 @@ class ClientDetail(Base):
         roles: list[str] = [],
     ) -> ClientDetail | None:
         """Fetch a database record via a client_id"""
-        if object_id == config.security.oauth_root_client_id:
-            return _get_root_client_detail(config, scopes=scopes, roles=roles)
+        root_client_id = get_client_id(db)
+        if object_id in (config.security.oauth_root_client_id, root_client_id):
+            return _get_root_client_detail(
+                config, scopes=scopes, roles=roles, client_id=root_client_id
+            )
         return super().get(db, object_id=object_id)
 
     def create_access_code_jwe(self, encryption_key: str) -> str:
@@ -139,6 +143,7 @@ def _get_root_client_detail(
     config: FidesConfig,
     scopes: list[str],
     roles: list[str],
+    client_id: str,
     encoding: str = "UTF-8",
 ) -> ClientDetail | None:
     """
@@ -149,7 +154,7 @@ def _get_root_client_detail(
 
     if scopes or roles:
         return ClientDetail(
-            id=config.security.oauth_root_client_id,
+            id=client_id,
             hashed_secret=config.security.oauth_root_client_secret_hash[0],
             salt=config.security.oauth_root_client_secret_hash[1].decode(encoding),
             scopes=scopes,
@@ -158,10 +163,27 @@ def _get_root_client_detail(
         )
 
     return ClientDetail(
-        id=config.security.oauth_root_client_id,
+        id=client_id,
         hashed_secret=config.security.oauth_root_client_secret_hash[0],
         salt=config.security.oauth_root_client_secret_hash[1].decode(encoding),
         scopes=DEFAULT_SCOPES,
         roles=DEFAULT_ROLES,
         systems=DEFAULT_SYSTEMS,
     )
+
+
+def get_client_id(db_session: Session) -> str:
+    client = ClientDetail.get_by(
+        db=db_session,
+        field="fides_key",
+        value=DEFAULT_OAUTH_CLIENT_KEY,
+    )
+    if not client:
+        client, _ = ClientDetail.create_client_and_secret(
+            db=db_session,
+            client_id_byte_length=CONFIG.security.oauth_client_id_length_bytes,
+            client_secret_byte_length=CONFIG.security.oauth_client_secret_length_bytes,
+            fides_key=DEFAULT_OAUTH_CLIENT_KEY,
+        )
+
+    return client.id
